+++ src/ui/icons.jsx
@@ -57,6 +57,7 @@ export const Icon = {
   text: () => <S><path d="M5 6V4h14v2M12 4v16M9 20h6" /></S>,
   size: () => <S><path d="M12 4v16M8 8l4-4 4 4M8 16l4 4 4-4" /></S>,
   align: () => <S><path d="M4 6h16M7 12h10M5 18h14" /></S>,
+  leading: () => <S><path d="M10 6h10M10 12h10M10 18h10" /><path d="M4 5v14M2 7l2-2 2 2M2 17l2 2 2-2" /></S>,
   front: () => <S><path d="M4 8l8-4 8 4-8 4-8-4z" /><path d="M4 14l8 4 8-4" /></S>,
   back: () => <S><path d="M4 16l8 4 8-4-8-4-8 4z" /><path d="M4 10l8-4 8 4" /></S>,
   mute: () => (